pipeman-master/requirements.txt
--index-url http://pypi-ni.lanni/simple/ --trusted-host pypi-ni.lanni
asgiref==3.7.2
beautifulsoup4==4.12.2
cachetools==4.2.4
certifi==2023.5.7
charset-normalizer==3.1.0
crispy-bootstrap4==2022.1
decorator==4.4.2
Django==4.2.1
django-betterforms==2.0.0
django-crispy-forms==2.0
django-impersonate==1.9.1
django-unicorn==0.51.1
django-vuexy==1.0.7
fontawesomefree==6.4.0
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
orjson==3.9.0
Pillow==9.5.0
python-gitlab==3.14.0
requests==2.31.0
requests-toolbelt==1.0.0
shortuuid==1.0.11
smmap==5.0.0
sorl-thumbnail==2023.3.10
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.6.2
urllib3==2.0.2
